// packages/shared/generated/serverTypes.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type ConversationAdminType = {
  _adminId: Scalars['String'];
  _id: Scalars['String'];
};

export type ConversationBasicType = {
  _conversationId: Scalars['String'];
  _id: Scalars['String'];
};

export type ConversationReturnType = {
  data: ConversationType;
  membersIds: Array<Scalars['String']>;
};

export type ConversationType = {
  _adminsIds: Array<ConversationAdminType>;
  _id: Scalars['String'];
  _membersId: Scalars['String'];
  _messagesId: Scalars['String'];
  name?: Maybe<Scalars['String']>;
};

export type FriendType = {
  _id: Scalars['String'];
  _userId: Scalars['String'];
};

export type FriendsBasicType = {
  _id: Scalars['String'];
  _userId: Scalars['String'];
};
// ...
export type UserType = {
  _id: Scalars['String'];
  conversations: Array<ConversationBasicType>;
  date: Scalars['String'];
  email: Scalars['String'];
  friends: FriendsBasicType;
  hash: Scalars['String'];
  name: Scalars['String'];
  salt: Scalars['String'];
  token: Scalars['String'];
};